Consolidated Reference processing where possible to Transformers
