+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/medicationorder/Dstu2MedicationOrderIncludesIcnMajig.java
@@ -1,7 +1,8 @@
 package gov.va.api.health.dataquery.service.controller.medicationorder;
 
+import static gov.va.api.health.dataquery.service.controller.Transformers.asReferenceId;
+
 import gov.va.api.health.argonaut.api.resources.MedicationOrder;
-import gov.va.api.health.dataquery.service.controller.Dstu2Transformers;
 import gov.va.api.health.dataquery.service.controller.IncludesIcnMajig;
 import gov.va.api.health.dstu2.api.bundle.AbstractEntry;
 import java.util.stream.Stream;
@@ -22,6 +23,6 @@ public class Dstu2MedicationOrderIncludesIcnMajig implements ResponseBodyAdvice<
           .type(MedicationOrder.class)
           .bundleType(MedicationOrder.Bundle.class)
           .extractResources(bundle -> bundle.entry().stream().map(AbstractEntry::resource))
-          .extractIcns(body -> Stream.ofNullable(Dstu2Transformers.asReferenceId(body.patient())))
+          .extractIcns(body -> Stream.ofNullable(asReferenceId(body.patient())))
           .build();
 }
