+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/TransformersTest.java
@@ -1,18 +1,29 @@
 package gov.va.api.health.dataquery.service.controller;
 
+import static gov.va.api.health.dataquery.service.controller.Transformers.asDatamartReference;
 import static gov.va.api.health.dataquery.service.controller.Transformers.asDateString;
 import static gov.va.api.health.dataquery.service.controller.Transformers.asDateTimeString;
+import static gov.va.api.health.dataquery.service.controller.Transformers.asReferenceId;
+import static gov.va.api.health.dataquery.service.controller.Transformers.convert;
+import static gov.va.api.health.dataquery.service.controller.Transformers.emptyToNull;
+import static gov.va.api.health.dataquery.service.controller.Transformers.ifPresent;
 import static gov.va.api.health.dataquery.service.controller.Transformers.isBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.parseInstant;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import gov.va.api.health.dataquery.service.controller.datamart.DatamartReference;
+import gov.va.api.health.dstu2.api.elements.Reference;
 import java.time.Instant;
 import java.time.LocalDate;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.Function;
 import org.junit.Test;
 
 public class TransformersTest {
+
   @Test
   public void allBlank() {
     assertThat(Transformers.allBlank()).isTrue();
@@ -20,6 +31,31 @@ public class TransformersTest {
     assertThat(Transformers.allBlank(null, "", " ")).isTrue();
     assertThat(Transformers.allBlank(null, 1, null, null)).isFalse();
     assertThat(Transformers.allBlank(1, "x", "z", 2.0)).isFalse();
+  }
+
+  @Test
+  public void asDatamartReferenceReturnsNullWhenRefIsEmpty() {
+    Reference ref = Reference.builder().build();
+    assertThat(asDatamartReference(ref)).isNull();
+  }
+
+  @Test
+  public void asDatamartReferenceReturnsNullWhenRefIsNotResourceSlashId() {
+    assertThat(asDatamartReference(Reference.builder().reference("no").build())).isNull();
+    assertThat(asDatamartReference(Reference.builder().reference("/no").build())).isNull();
+    assertThat(asDatamartReference(Reference.builder().reference("no/").build())).isNull();
+    assertThat(asDatamartReference(Reference.builder().reference("no/ ").build())).isNull();
+    assertThat(asDatamartReference(Reference.builder().reference("no/no").build())).isNull();
+    assertThat(asDatamartReference(Reference.builder().reference("whatever/NO/").build())).isNull();
+  }
+
+  @Test
+  public void asDatamartReferenceReturnsRef() {
+    Reference ref = Reference.builder().display("WAT").reference("Patient/123V456").build();
+    DatamartReference dataRef = asDatamartReference(ref);
+    assertThat(dataRef.type().get()).isEqualTo("Patient");
+    assertThat(dataRef.reference().get()).isEqualTo("123V456");
+    assertThat(dataRef.display().get()).isEqualTo("WAT");
   }
 
   @Test
@@ -77,6 +113,57 @@ public class TransformersTest {
   }
 
   @Test
+  public void asReferenceIdReturnsNullWhenRefIsEmpty() {
+    Reference ref = Reference.builder().build();
+    assertThat(asReferenceId(ref)).isNull();
+  }
+
+  @Test
+  public void asReferenceIdReturnsRefId() {
+    Reference ref = Reference.builder().display("WAT").reference("Patient/123V456").build();
+    assertThat(asReferenceId(ref)).isEqualTo("123V456");
+  }
+
+  @Test
+  public void convertReturnsConvertedWhenItemIsPopulated() {
+    Function<Integer, String> tx = o -> "x" + o;
+    assertThat(convert(1, tx)).isEqualTo("x1");
+  }
+
+  @Test
+  public void convertReturnsNullWhenItemIsNull() {
+    Function<String, String> tx = o -> "x" + o;
+    assertThat(convert(null, tx)).isNull();
+  }
+
+  @Test
+  public void emptyToNullReturnsNonNullEntries() {
+    assertThat(emptyToNull(Arrays.asList("a", null, "b", null))).isEqualTo(List.of("a", "b"));
+  }
+
+  @Test
+  public void emptyToNullReturnsNullIfEmpty() {
+    assertThat(emptyToNull(List.of())).isNull();
+  }
+
+  @Test
+  public void emptyToNullReturnsNullIfNull() {
+    assertThat(emptyToNull(null)).isNull();
+  }
+
+  @Test
+  public void ifPresentReturnsExtractWhenObjectIsNull() {
+    Function<Object, String> extract = (o) -> "x" + o;
+    assertThat(ifPresent("a", extract)).isEqualTo("xa");
+  }
+
+  @Test
+  public void ifPresentReturnsNullWhenObjectIsNull() {
+    Function<Object, String> extract = (o) -> "x" + o;
+    assertThat(ifPresent(null, extract)).isNull();
+  }
+
+  @Test
   public void isBlankCollection() {
     assertThat(isBlank(List.of())).isTrue();
     assertThat(isBlank(List.of("x"))).isFalse();
@@ -86,5 +173,20 @@ public class TransformersTest {
   public void isBlankMap() {
     assertThat(isBlank(Map.of())).isTrue();
     assertThat(isBlank(Map.of("x", "y"))).isFalse();
+  }
+
+  @Test
+  public void isBlankOptional() {
+    assertThat(isBlank(Optional.empty())).isTrue();
+    assertThat(isBlank(Optional.of(""))).isTrue();
+    assertThat(isBlank(Optional.of("x"))).isFalse();
+  }
+
+  @Test
+  public void parseInstantParsesWhenPossible() {
+    Instant expected = Instant.parse("2005-01-12T07:57:00Z");
+    assertThat(parseInstant("2005-01-12T07:57:00Z")).isEqualTo(expected);
+    assertThat(parseInstant("2005-01-12T07:57:00")).isEqualTo(expected);
+    assertThat(parseInstant("nope")).isNull();
   }
 }
