+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/diagnosticreport/DiagnosticReportSamples.java
@@ -1,7 +1,7 @@
 package gov.va.api.health.dataquery.service.controller.diagnosticreport;
 
 import static gov.va.api.health.autoconfig.configuration.JacksonConfig.createMapper;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.parseInstant;
+import static gov.va.api.health.dataquery.service.controller.Transformers.parseInstant;
 import static java.util.Arrays.asList;
 
 import gov.va.api.health.argonaut.api.resources.DiagnosticReport;
